requirements.txt
flask
google-generativeai
gunicorn  # Để chạy production trên Render
